add default case in workflows.go (#90)
* change kustomize to default in workflows.go

update the image tag for manifes. For kustomize and helm we use a production override which has the image modified however, we do not do this for regular k8s manifests. This PR changes that.

* Add Karen's suggestion

Returning the kustomize case and having a case directly for the manifests pointing to the manifests folder

* add missing bracket